pyproject.toml
[project]
name = "hackai-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "dotenv>=0.9.9",
    "fitz>=0.0.1.dev2",
    "frontend>=0.0.3",
    "pillow>=12.1.1",
    "presidio-analyzer>=2.2.361",
    "presidio-anonymizer>=2.2.361",
    "pymupdf>=1.27.1",
    "pytesseract>=0.3.13",
    "pyutilities>=2.1.2",
    "utility>=1.0",
]
